+++ gems/quilt_rails/lib/generators/sewing_kit/install_generator.rb
@@ -6,16 +6,29 @@ module SewingKit
 
     desc "This generator creates a sewing-kit config file."
 
-    def create_config
-      config_path = "config/sewing-kit.config.ts"
+    def initialize(args, *options)
+      @application_name = Rails.application.class.module_parent.to_s.underscore
+      super(args, *options)
+    end
 
-      if File.exist?(config_path)
-        say "Sewing kit config already exists"
-      else
-        copy_file "sewing-kit.config.ts", config_path
+    def create_package_json
+      package_json_path = "package.json"
 
-        say "Sewing kit config"
-      end
+      copy_file(package_json_path)
+      gsub_file(package_json_path, "\${application_name}", @application_name)
+    end
+
+    def create_sk_config
+      sk_config_path = "config/sewing-kit.config.ts"
+
+      copy_file("sewing-kit.config.ts", sk_config_path)
+      gsub_file(sk_config_path, "\${application_name}", @application_name)
+    end
+
+    def create_config_files
+      copy_file(".editorconfig")
+      copy_file(".eslintignore")
+      copy_file(".prettierignore")
     end
   end
 end