+++ packages/dates/src/is-less-than-one-year-away.ts
@@ -1,0 +1,8 @@
+import {isFutureDate} from './is-future-date';
+import {TimeUnit} from './constants';
+
+export function isLessThanOneYearAway(date: Date, now = new Date()) {
+  return (
+    isFutureDate(date, now) && date.getTime() - now.getTime() < TimeUnit.Year
+  );
+}