+++ packages/react-server/src/webpack-plugin/test/webpack-plugin.test.ts
@@ -2,13 +2,13 @@ import path from 'path';
 
 import webpack, {Compiler} from 'webpack';
 
-import {HEADER, Options} from '../react-server-webpack-plugin';
+import {HEADER, Options} from '../shared';
 
 import {withWorkspace} from './utilities/workspace';
 
 const BUILD_TIMEOUT = 10000;
 
-describe('react-server-webpack-plugin', () => {
+describe('webpack-plugin', () => {
   describe('node', () => {
     it(
       'generates the server and client entrypoints when the virtual server & client modules are present',
@@ -225,12 +225,91 @@ describe('react-server-webpack-plugin', () => {
       BUILD_TIMEOUT,
     );
   });
+
+  describe('error component', () => {
+    it(
+      'imports the Error component in the server when error.js exists',
+      async () => {
+        const name = 'rails-includes-error-component';
+        const basePath = './app/ui';
+
+        await withWorkspace(name, async ({workspace}) => {
+          await workspace.write(`${basePath}/index.js`, BASIC_JS_MODULE);
+          await workspace.write(
+            `${basePath}/error.js`,
+            BASIC_JS_ERROR_COMPOENT,
+          );
+
+          await workspace.write(
+            'webpack.config.js',
+            createWebpackConfig({basePath}),
+          );
+
+          const [serverResults] = await runBuild(name);
+          const serverModule = getModule(serverResults, 'app/ui/server');
+
+          expect(serverModule.source).toMatch("import Error from 'error';");
+        });
+      },
+      BUILD_TIMEOUT,
+    );
+
+    it(
+      'does not import the Error component in the server when error.js does not exists',
+      async () => {
+        const name = 'rails-doesnt-include-error-component';
+        const basePath = './app/ui';
+
+        await withWorkspace(name, async ({workspace}) => {
+          await workspace.write(`${basePath}/index.js`, BASIC_JS_MODULE);
+
+          await workspace.write(
+            'webpack.config.js',
+            createWebpackConfig({basePath}),
+          );
+
+          const [serverResults] = await runBuild(name);
+          const serverModule = getModule(serverResults, 'app/ui/server');
+
+          expect(serverModule.source).not.toMatch("import Error from 'error';");
+        });
+      },
+      BUILD_TIMEOUT,
+    );
+
+    it(
+      'includes the Error component in the server when an error folder exists',
+      async () => {
+        const name = 'rails-includes-error-component-folder';
+        const basePath = './app/ui';
+
+        await withWorkspace(name, async ({workspace}) => {
+          await workspace.write(`${basePath}/index.js`, BASIC_JS_MODULE);
+          await workspace.write(
+            `${basePath}/error/index.js`,
+            BASIC_JS_ERROR_COMPOENT,
+          );
+
+          await workspace.write(
+            'webpack.config.js',
+            createWebpackConfig({basePath}),
+          );
+
+          const [serverResults] = await runBuild(name);
+          const serverModule = getModule(serverResults, 'app/ui/server');
+
+          expect(serverModule.source).toMatch("import Error from 'error';");
+        });
+      },
+      BUILD_TIMEOUT,
+    );
+  });
 });
 
 function runBuild(configPath: string): Promise<any[]> {
   return new Promise((resolve, reject) => {
     const pathFromRoot = path.resolve(
-      './packages/react-server-webpack-plugin/src/test/fixtures',
+      './packages/react-server/src/webpack-plugin/test/fixtures',
       configPath,
     );
 
@@ -290,13 +369,17 @@ const BASIC_JS_MODULE = `module.exports = () => {
   return 'I am totally a react component';
 };`;
 
+const BASIC_JS_ERROR_COMPOENT = `module.exports = () => {
+  return 'I am totally an Error component';
+};`;
+
 const createWebpackConfig = (
   {basePath, port, host, assetPrefix}: Options = {
     basePath: '.',
   },
 ) => `
 const path = require('path');
-const {ReactServerPlugin} = require('../../../react-server-webpack-plugin');
+const {ReactServerPlugin} = require('../../../webpack-plugin');
 
 const universal = {
   mode: 'production',