+++ packages/enzyme-utilities/CHANGELOG.md
@@ -7,6 +7,12 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [2.1.12] - 2020-05-29
+
+### Fixed
+
+- Remove cast for `act` following update to `@types/react-dom`
+
 ## [2.1.11] - 2020-05-13
 
 - `trigger` update the wrapper even if the promise is not resolved. [[#1439](https://github.com/Shopify/quilt/pull/1439)]